requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dotenv==0.0.5
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
PyJWT==2.8.0
python-dotenv==1.0.0
python-telegram-bot==20.5
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
